requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SpeechRecognition==3.10.1
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
